[WGSL] Using UINT_MAX for location triggers a RELEASE_ASSERT https://bugs.webkit.org/show_bug.cgi?id=286758 rdar://143806211
Reviewed by Mike Wyrzykowski.

This is very similar to 289563@main, but applies to the `location` attribute instead of
the `binding` attribute. In short, the vlaue of attribute is used as the key to a HashMap,
but the key had type `unsigned` and `UINT_MAX` was used as the deleted value, so using
`@location(UINT_MAX)` triggered a release assert in the HashMap.

* Source/WebGPU/WGSL/AttributeValidator.cpp:

Canonical link: https://commits.webkit.org/289729@main