[GTK][WPE] Crash loading https://webassembly.sh/ in CPU rendering mode https://bugs.webkit.org/show_bug.cgi?id=286674
Reviewed by Carlos Garcia Campos.

Fix the check to paint the contents on the canvas when its ImageBuffer hasn't
been created yet: if we do need to paint the transparent black rectangle, do
not call surfaceBufferToImageBuffer() because that will trigger the creation
of the ImageBuffer, potentially changing the canvas into an accelerated one in
the middle of the paint, which will cause a crash. Paint the transparent black
rectangle using fillRect() instead.

* Source/WebCore/html/HTMLCanvasElement.cpp:
(WebCore::HTMLCanvasElement::paint):

Canonical link: https://commits.webkit.org/289775@main