Resuming an interrupted `git webkit pr` produces a commit message without links rdar://125444927 https://bugs.webkit.org/show_bug.cgi?id=271279
Reviewed by Jonathan Bedard.

When creating a commit in program.commit or program.pull_request, we shouldn't
set env variables if the bug title and urls are null.

If these are empty, we check for the variables stored in the branch config.

* Tools/Scripts/hooks/prepare-commit-msg:
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/commit.py:
(Commit.main):
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/pull_request.py:
(PullRequest.create_commit):

Canonical link: https://commits.webkit.org/289804@main