Introduce the StageMode attribute to Model https://bugs.webkit.org/show_bug.cgi?id=285899 rdar://142566328
Reviewed by Ada Chan.

This is the first of three PRs used to introduce the stagemode attribute to model views, which will
be used to define and control the interaction of a model. The stagemode attribute takes a string and
determines how the model behaves based on user's gesture input on the view containing the model. In
particular, we currently support 2 types of stagemode behaviors: orbit, which will allow users to rotate
the model, and none, which will disable model interaction. Subsequent PRs will be made to use this
attribute to define model interaction.

* LayoutTests/model-element/model-element-stagemode-expected.txt: Added.
* LayoutTests/model-element/model-element-stagemode.html: Added.
* Source/WebCore/Modules/model-element/HTMLModelElement.cpp:
(WebCore::HTMLModelElement::createModelPlayer):
Updates the stagemode attribute when the source for the model element updates
(WebCore::HTMLModelElement::attributeChanged):
Updates the stagemode attribute when the attribute changes
(WebCore::HTMLModelElement::stageMode const):
Returns the enum corresponding to the stagemode attribute as defined by the user-defined string
(WebCore::HTMLModelElement::updateStageMode):
Updates the internal variable the model player uses to track the stagemode attribute
* Source/WebCore/Modules/model-element/HTMLModelElement.h:
* Source/WebCore/Modules/model-element/HTMLModelElement.idl:
* Source/WebCore/Modules/model-element/ModelPlayer.cpp:
(WebCore::ModelPlayer::setStageMode):
Virtual function to update the player based on the stagemode attribute
* Source/WebCore/Modules/model-element/ModelPlayer.h:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/html/HTMLAttributeNames.in:
* Source/WebCore/page/StageModeOperations.h: Added.
* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.h:
* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.messages.in:
* Source/WebKit/ModelProcess/cocoa/ModelProcessModelPlayerProxy.mm:
(WebKit::ModelProcessModelPlayerProxy::setStageMode):
Updates the internal variable correpsonding to the stagemode attribute
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/WebProcess/Model/ModelProcessModelPlayer.cpp:
(WebKit::ModelProcessModelPlayer::setStageMode):
Updates the internal variable correpsonding to the stagemode attribute
* Source/WebKit/WebProcess/Model/ModelProcessModelPlayer.h:

Canonical link: https://commits.webkit.org/289830@main