Crash in com.apple.WebKit.WebContent at com.apple.WebKit: WebKit::WebExtensionAPIEvent::removeAllListeners https://bugs.webkit.org/show_bug.cgi?id=287095 rdar://144211832
Reviewed by Timothy Hatcher and Brian Weinstein.

This patch is a speculative fix for a crash introduced by 289222@main, as I
have been unable to reproduce the problem.

The crash is almost certainly happening because we're dereferencing a nullptr,
so the fix is to simply add null checks prior to dereferencing.

* Source/WebKit/WebProcess/Extensions/API/Cocoa/WebExtensionAPIEventCocoa.mm:
(WebKit::WebExtensionAPIEvent::addListener):
(WebKit::WebExtensionAPIEvent::removeListener):
(WebKit::WebExtensionAPIEvent::removeAllListeners):

Canonical link: https://commits.webkit.org/289898@main