Fix pause in popups, create target when WebPage is initialized
There are a few ways a page can be created and navigated for the first time:
  - context.newPage
  - click navigation in a new window
    -- opener/noopener
  - window.open() navigation
    -- opener/noopener
In case of popup windows the creation starts with a sync `WebPageProxy::createNewPage` message from `WebChromeClient::createWindow` to the `WebPageProxy` and it may end up creating `WebPage` in the same web process as the opener. The page in that case should be paused on start to give Playwright client a chance to configure the new page before returning control to the caller (e.g. `window.open()`).

Previously the `noopener` page was created in a new process and wouldn't have a related page. By the time `shouldPauseInInspectorWhenShown` was called either:
  - the web process already existed and was blocked on sending sync message `WebPageProxy::createNewPage` from `WebChromeClient::createWindow` or
  - a new process would be launched later and the new popup `WebPage` would be initialized in it
In both cases it was safe to create new `InspectorTargetProxy` and send `Messages::WebPage::ConnectInspector` as the message would either be buffered up (because IPC was blocked) or it would be sucessfully dispatched to just created (and attached to) new page.

After the upstream change 290596@main in `WebPageProxy.cpp` the popup page with noopener could be created in already running process, but there were two problems:
   - `setOpenedByDOM` would be called in a callback inside `WebPageProxy::createNewPage` after `WebPageProxy::initializeWebPage` called `shouldPauseInInspectorWhenShown` which would result in non paused `WebPage`
   - `InspectorTargetProxy` would be created right away inside `WebPageInspectorController::init` as there was already running web process and it would immediately send `Messages::WebPage::ConnectInspector` to the process. The message would be dispatched before `WebPage` was actually initialized there.

This PR ensures that:
  - `InspectorTargetProxy` is created and connected to the WebPage`s inspector only after `WebPage` has been created in the web process;
  - `shouldPauseInInspectorWhenShown` detects popups for pause by looking whether `windowFeautures` are present instead of relying on `isOpenedByDOM` which may be not set yet.

It also reverts our revert of the upstream change in 290596@main in `WebPageProxy.cpp`.