[Win] Regression - crashing while drawing text blob https://bugs.webkit.org/show_bug.cgi?id=300412
Reviewed by Fujii Hironori.

In 300818@main we assumed that the only possible GraphicsContext for which we
replay a display list is GraphicsContextSkia, which is true for GTK and
WPE ports that don't use the GPU process yet for DOM rendering, but not
for win port. This patch adds a new DisplayList::Recorder::DrawGlyphsMode
for skia that uses a new DrawTextBlob display list item, leaving
DrawGlyphs item as it was before 300818@main. This new mode is only set
then recording for a GraphicsContext that has a platform context.

Canonical link: https://commits.webkit.org/301321@main