Unfork downstream change introduced to fix https://github.com/microsoft/playwright/issues/37766
The crash was fixed upstream by 302279@main

The upstream crash is very similiar to the first attempt to fix a crash
in page/page-history.spec.ts:143. Then an user reported issue/37766, which can
be reproduced when running page.reload multiple times in a row. I
confirmed neither of these two tests are now crashing in WebKit main.